Fixes a bug in the Base::populateTimeLogFromPostData()
If a user posted another tag alongside a meta tags things would not get translated correctly.
By iterating over each posted field individually to see if we are dealing with a plain tag or a meta tag and acting accordingly, things are now handled as they should be.

This allows for other tags (like comment or group tags) to be added to meta tag for inclusion for persistence.